pyproject.toml
[project]
name = "google-merchant-feed-service"
version = "0.1.0"
description = "Google Merchant feed synchronization microservice"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
  "fastapi>=0.110",
  "uvicorn>=0.27",
  "psycopg2-binary>=2.9",
  "gspread>=6.0",
  "google-auth>=2.29",
  "google-api-python-client>=2.120",
  "python-dotenv>=1.0",
  "redis>=5.0",
  "apscheduler>=3.10",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "httpx>=0.27",
  "pre-commit>=3.7",
  "ruff>=0.0.300"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "I", "C90"]  # example, can expand as needed

[tool.ruff.lint]
select = ["E", "F", "I"]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
]
